Move tests from test data file
