+++ src/lib/assets/rbac.test.ts
@@ -4,34 +4,285 @@ import { clusterRole, clusterRoleBinding, storeRole, serviceAccount, storeRoleBi
 import { it, describe, expect, beforeEach, jest } from "@jest/globals";
 import { V1PolicyRule as PolicyRule } from "@kubernetes/client-node";
 import fs from "fs";
-import { mockCapabilities } from "./defaultTestObjects";
+import {
+  capabilitiesWithDuplicates,
+  capabilitiesWithFinalize,
+  capabilitiesWithLongKey,
+  capabilitiesWithShortKey,
+  mockCapabilities,
+} from "./defaultTestObjects";
+import { ModuleConfig } from "../types";
+import { WebhookIgnore } from "../k8s";
+import { kind } from "kubernetes-fluent-client";
+import * as helpers from "../helpers";
 
+describe("RBAC generation", () => {
+  beforeEach(() => {
+    jest.clearAllMocks();
+    const mockPackageJsonRBAC = {};
+
+    jest.spyOn(fs, "readFileSync").mockImplementation((path: unknown) => {
+      if (typeof path === "string" && path.includes("package.json")) {
+        return JSON.stringify({ rbac: mockPackageJsonRBAC });
+      }
+      return "{}";
+    });
+  });
+
+  it("should generate correct ClusterRole rules in scoped mode", () => {
+    const result = clusterRole("test-role", mockCapabilities, "scoped", []);
+
+    expect(result.rules).toEqual([
+      {
+        apiGroups: ["pepr.dev"],
+        resources: ["peprstores"],
+        verbs: ["create", "get", "patch", "watch"],
+      },
+      {
+        apiGroups: ["apiextensions.k8s.io"],
+        resources: ["customresourcedefinitions"],
+        verbs: ["patch", "create"],
+      },
+      {
+        apiGroups: [""],
+        resources: ["namespaces"],
+        verbs: ["watch"],
+      },
+      {
+        apiGroups: [""],
+        resources: ["configmaps"],
+        verbs: ["watch"],
+      },
+    ]);
+  });
+
+  it("should generate a ClusterRole with wildcard rules when not in scoped mode", () => {
+    const expectedWildcardRules = [
+      {
+        apiGroups: ["*"],
+        resources: ["*"],
+        verbs: ["create", "delete", "get", "list", "patch", "update", "watch"],
+      },
+    ];
+
+    const result = clusterRole("test-role", mockCapabilities, "admin", []);
+
+    expect(result.rules).toEqual(expectedWildcardRules);
+  });
+
+  it("should return an empty rules array when capabilities are empty in scoped mode", () => {
+    const result = clusterRole("test-role", [], "scoped", []);
+
+    expect(result.rules).toEqual([]);
+  });
+
+  it("should include finalize verbs if isFinalize is true in scoped mode", () => {
+    const result = clusterRole(
+      "test-role",
+      capabilitiesWithFinalize,
+      "scoped",
+      capabilitiesWithFinalize.flatMap(c => c.rbac).filter((rule): rule is PolicyRule => rule !== undefined),
+    );
+
+    expect(result.rules).toEqual([
+      {
+        apiGroups: ["pepr.dev"],
+        resources: ["peprstores"],
+        verbs: ["patch"],
+      },
+      {
+        apiGroups: ["apiextensions.k8s.io"],
+        resources: ["customresourcedefinitions"],
+        verbs: ["patch", "create"],
+      },
+    ]);
+  });
+
+  it("should deduplicate verbs and resources in rules", () => {
+    const result = clusterRole(
+      "test-role",
+      capabilitiesWithDuplicates,
+      "scoped",
+      capabilitiesWithDuplicates.flatMap(c => c.rbac).filter((rule): rule is PolicyRule => rule !== undefined),
+    );
+
+    // Filter out only the rules for 'pepr.dev' and 'peprstores'
+    const filteredRules = result.rules?.filter(
+      rule => rule.apiGroups?.includes("pepr.dev") && rule.resources?.includes("peprstores"),
+    );
+
+    expect(filteredRules).toEqual([
+      {
+        apiGroups: ["pepr.dev"],
+        resources: ["peprstores"],
+        verbs: ["create", "get", "patch", "watch"],
+      },
+    ]);
+  });
+});
+describe("clusterRole", () => {
+  // Mocking the readRBACFromPackageJson function to return null
+  jest.mock("./rbac", () => ({
+    ...(jest.requireActual("./rbac") as object),
+    readRBACFromPackageJson: jest.fn(() => null),
+  }));
+
+  // Mocking createRBACMap to isolate the behavior of clusterRole function
+  jest.mock("../helpers", () => ({
+    ...(jest.requireActual("../helpers") as object),
+    createRBACMap: jest.fn(),
+  }));
+
+  beforeEach(() => {
+    jest.clearAllMocks();
+    jest.restoreAllMocks();
+  });
+
+  it("should handle keys with less than 3 segments and set group to an empty string", () => {
+    jest.spyOn(helpers, "createRBACMap").mockReturnValue({
+      nodes: {
+        plural: "nodes",
+        verbs: ["get"],
+      },
+    });
+
+    const result = clusterRole(
+      "test-role",
+      capabilitiesWithShortKey,
+      "scoped",
+      capabilitiesWithShortKey.flatMap(c => c.rbac).filter((rule): rule is PolicyRule => rule !== undefined),
+    );
+
+    expect(result.rules).toEqual([
+      {
+        apiGroups: [""],
+        resources: ["nodes"],
+        verbs: ["get"],
+      },
+    ]);
+  });
+
+  it("should handle keys with 3 or more segments and set group correctly", () => {
+    jest.spyOn(helpers, "createRBACMap").mockReturnValue({
+      "apps/v1/deployments": {
+        plural: "deployments",
+        verbs: ["create"],
+      },
+    });
+
+    const result = clusterRole(
+      "test-role",
+      capabilitiesWithLongKey,
+      "scoped",
+      capabilitiesWithLongKey.flatMap(c => c.rbac).filter((rule): rule is PolicyRule => rule !== undefined),
+    );
+
+    expect(result.rules).toEqual([
+      {
+        apiGroups: ["apps"],
+        resources: ["deployments"],
+        verbs: ["create"],
+      },
+    ]);
+  });
+
+  it("should handle non-array custom RBAC by defaulting to an empty array", () => {
+    // Mock readRBACFromPackageJson to return a non-array value
+    jest.spyOn(fs, "readFileSync").mockImplementation(() => {
+      return JSON.stringify({
+        pepr: {
+          rbac: "not-an-array", // Simulate invalid RBAC structure
+        },
+      });
+    });
+
+    const result = clusterRole(
+      "test-role",
+      mockCapabilities,
+      "scoped",
+      mockCapabilities.flatMap(c => c.rbac).filter((rule): rule is PolicyRule => rule !== undefined),
+    );
+
+    // The result should only contain rules from the capabilities, not from the invalid custom RBAC
+    expect(result.rules).toEqual([
+      {
+        apiGroups: ["pepr.dev"],
+        resources: ["peprstores"],
+        verbs: ["create", "get", "patch", "watch"],
+      },
+      {
+        apiGroups: ["apiextensions.k8s.io"],
+        resources: ["customresourcedefinitions"],
+        verbs: ["patch", "create"],
+      },
+      {
+        apiGroups: [""],
+        resources: ["namespaces"],
+        verbs: ["watch"],
+      },
+      {
+        apiGroups: [""],
+        resources: ["configmaps"],
+        verbs: ["watch"],
+      },
+    ]);
+  });
+
+  it("should default to an empty verbs array if rule.verbs is undefined", () => {
+    // Simulate a custom RBAC rule with empty verbs
+    const customRbacWithNoVerbs: PolicyRule[] = [
+      {
+        apiGroups: ["pepr.dev"],
+        resources: ["customresources"],
+        verbs: [], // Set verbs to an empty array to satisfy the V1PolicyRule type
+      },
+    ];
+
+    jest.spyOn(fs, "readFileSync").mockImplementation(() => {
+      return JSON.stringify({
+        pepr: {
+          rbac: customRbacWithNoVerbs,
+        },
+      });
+    });
+
+    const result = clusterRole("test-role", mockCapabilities, "scoped", customRbacWithNoVerbs);
+
+    // Check that the verbs array is empty for the custom RBAC rule
+    expect(result.rules).toContainEqual({
+      apiGroups: ["pepr.dev"],
+      resources: ["customresources"],
+      verbs: [],
+    });
+  });
+});
 describe("RBAC generation with mocked package.json", () => {
   beforeEach(() => {
     jest.clearAllMocks();
     jest.spyOn(fs, "readFileSync").mockImplementation((path: unknown) => {
       if (typeof path === "string" && path.includes("package.json")) {
-        return JSON.stringify({
-          pepr: {
-            rbac: [
-              {
-                apiGroups: ["pepr.dev"],
-                resources: ["pods"],
-                verbs: ["get", "list"],
-              },
-              {
-                apiGroups: ["pepr.dev"],
-                resources: ["pods"],
-                verbs: ["list"],
-              },
-              {
-                apiGroups: ["apps"],
-                resources: ["deployments"],
-                verbs: ["create", "delete"],
-              },
-            ],
-          },
-        });
+        const moduleConfig: ModuleConfig = {
+          uuid: "some-uuid",
+          alwaysIgnore: jest.fn() as WebhookIgnore,
+          rbac: [
+            {
+              apiGroups: ["pepr.dev"],
+              resources: ["pods"],
+              verbs: ["get", "list"],
+            },
+            {
+              apiGroups: ["pepr.dev"],
+              resources: ["pods"],
+              verbs: ["list"],
+            },
+            {
+              apiGroups: ["apps"],
+              resources: ["deployments"],
+              verbs: ["create", "delete"],
+            },
+          ],
+        };
+        return JSON.stringify({ pepr: moduleConfig });
       }
       return "{}";
     });
@@ -60,7 +311,7 @@ describe("RBAC generation with mocked package.json", () => {
 describe("clusterRoleBinding", () => {
   it("should create a ClusterRoleBinding with the specified name", () => {
     const roleName = "test-cluster-role";
-    const expectedClusterRoleBinding = {
+    const expectedClusterRoleBinding: kind.ClusterRoleBinding = {
       apiVersion: "rbac.authorization.k8s.io/v1",
       kind: "ClusterRoleBinding",
       metadata: { name: roleName },
@@ -87,7 +338,7 @@ describe("clusterRoleBinding", () => {
 describe("serviceAccount", () => {
   it("should create a ServiceAccount with the specified name", () => {
     const accountName = "test-service-account";
-    const expectedServiceAccount = {
+    const expectedServiceAccount: kind.ServiceAccount = {
       apiVersion: "v1",
       kind: "ServiceAccount",
       metadata: {
@@ -105,7 +356,7 @@ describe("serviceAccount", () => {
 describe("storeRole", () => {
   it("should create a Role for managing peprstores with the specified name", () => {
     const roleName = "test-role";
-    const expectedRole = {
+    const expectedRole: kind.Role = {
       apiVersion: "rbac.authorization.k8s.io/v1",
       kind: "Role",
       metadata: {
@@ -131,7 +382,7 @@ describe("storeRole", () => {
 describe("storeRoleBinding", () => {
   it("should create a RoleBinding for the specified Role", () => {
     const roleName = "test-role";
-    const expectedRoleBinding = {
+    const expectedRoleBinding: kind.RoleBinding = {
       apiVersion: "rbac.authorization.k8s.io/v1",
       kind: "RoleBinding",
       metadata: {
